backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
pymysql==1.1.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
cryptography==44.0.1
